### 2023-01-14 #5080 @editorialbot
[PRE REVIEW]: DAPPER: Data Assimilation with Python: a Package for Experimental Research
**Submitting author:** @patnr (<a href="http://orcid.org/0000-0001-5073-6786">Patrick Nima Raanes</a>)
**Repository:** https://github.com/nansencenter/DAPPER
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** v1.4.0
