### 2023-01-14 #5080 @patnr
Thank you `arfon` for beginning the process, and the information.

If it's any help, a similar library is https://github.com/cgrudz/DataAssimilationBenchmarks.jl
with pre-review  #3976 and review #4129

The assigned editor was [taless474](https://github.com/taless474) while the assigned reviewers were [peanutfun](https://github.com/peanutfun) and [tmigot](https://github.com/tmigot). Among these `peanutfun` is perhaps particularly relevant since he knows about the subject and python. Please let me know if/when I should propose more reviewers.
