### 2023-03-29 8f49e54
Fix plots in Jupyter
