+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@ exclude: '^$|settings|scripts|docs'
 
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v3.3.0
+    rev: v4.4.0
     hooks:
     -   id: trailing-whitespace
         exclude: |
@@ -30,7 +30,7 @@ repos:
         args: ['--fix=no']
 
 -   repo: https://github.com/flakeheaven/flakeheaven
-    rev: 3.0.0
+    rev: 3.2.1
     hooks:
     -   id: flakeheaven
         additional_dependencies: [
@@ -43,7 +43,7 @@ repos:
         ]
 
 -   repo: https://github.com/timothycrosley/isort
-    rev: 5.6.4
+    rev: 5.12.0
     hooks:
     -   id: isort
         exclude: |
