+++ dapper/dpr_config.py
@@ -66,8 +66,7 @@ for d in rc.dirs:
 # Otherwise, warnings are thrown on every occurence of plt.pause
 # (though not plot_pause), and (who knows) maybe errors too.
 # Also, the assimilation slows down, even though nothing is shown.
-LP = rc.liveplotting
-if LP and not is_using_interactive_backend():
+if rc.liveplotting and not is_using_interactive_backend():
     # Note: plot_pause could maybe be adapted to also work for
     # "inline" backend (which is not striclty interactive), but I think
     # this would be buggy, and is incompatible with a "stop" button.
@@ -75,5 +74,4 @@ if LP and not is_using_interactive_backend():
           "in your dpr_config.yaml,",
           "but this is not supported by the current matplotlib backend:",
           f"{mpl.get_backend()}. To enable it, try using another backend.\n")
-    LP = False
-rc.liveplotting = LP
+    rc.liveplotting = False
