### 2023-03-30 31e6c5b
Make replays work with mpl inline (Colab)
