+++ dapper/__init__.py
@@ -4,7 +4,7 @@
 .. include:: ./README.md
 """
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"
 
 # A parsimonious list of imports used in the examples
 from .dpr_config import rc
