+++ dapper/tools/liveplotting.py
@@ -1325,7 +1325,8 @@ def spatial2d(
             #  - ind2sub returns (iy,ix), while plot takes (ix,iy) => reverse.
 
             if ko is not None and obs_inds is not None:
-                lh[0] = ax_12.plot(*ind2sub(obs_inds(ko))[::-1], 'k.', ms=1, zorder=5)[0]
+                lh[0] = ax_12.plot(*ind2sub(obs_inds(ko))[::-1],
+                                   'k.', ms=1, zorder=5)[0]
 
             text_t.set_text(format_time(k, ko, t))
 
