### 2023-05-18 fbc5f9e
Expand time-dep observation capabilities. Add example
