+++ dapper/tools/liveplotting.py
@@ -692,6 +692,13 @@ def sliding_marginals(
         p = DotDict(**{
             kw: kwargs.get(kw, val) for kw, val in params_orig.items()})
 
+        # No ndarray
+        p.dims = list(p.dims)
+
+        # Chose marginal dims to plot
+        if not p.dims:
+            p.dims = linspace_int(xx.shape[-1], min(10, xx.shape[-1]))
+
         # Lag settings:
         T_lag, K_lag, a_lag = validate_lag(p.Tplot, tseq)
         K_plot = comp_K_plot(K_lag, a_lag, plot_u)
@@ -700,31 +707,14 @@ def sliding_marginals(
         if has_w:
             K_plot += a_lag
 
-        # Chose marginal dims to plot
-        p.dims = list(p.dims)  # no ndarray
-        if not p.dims:
-            Nx      = min(10, xx.shape[-1])
-            DimsX   = linspace_int(xx.shape[-1], Nx)
-        else:
-            Nx      = len(p.dims)
-            DimsX   = p.dims
-        # Pre-process obs dimensions
-        # Rm inds of obs if not in DimsX
-        iiY   = [i for i, m in enumerate(p.obs_inds) if m in DimsX]
-        # Rm obs_inds    if not in DimsX
-        DimsY = [m for i, m in enumerate(p.obs_inds) if m in DimsX]
-        # Get dim (within y) of each x
-        DimsY = [DimsY.index(m) if m in DimsY else None for m in DimsX]
-        Ny    = len(iiY)
-
         # Set up figure, axes
-        fig, axs = place.freshfig(fignum, figsize=(5, 7), nrows=Nx, sharex=True)
-        if Nx == 1:
-            axs = [axs]
+        fig, axs = place.freshfig(fignum, figsize=(5, 7), squeeze=False,
+                                  nrows=len(p.dims), sharex=True)
+        axs = axs.reshape(len(p.dims))
 
         # Tune plots
         axs[0].set_title("Marginal time series")
-        for ix, (m, ax) in enumerate(zip(DimsX, axs)):
+        for ix, (m, ax) in enumerate(zip(p.dims, axs)):
             # ax.set_ylim(*viz.stretch(*viz.xtrema(xx[:, m]), 1/p.zoomy))
             if not p.labels:
                 ax.set_ylabel("$x_{%d}$" % m)
@@ -742,27 +732,27 @@ def sliding_marginals(
         if True:
             d.t  = RollingArray((K_plot,))
         if True:
-            d.x  = RollingArray((K_plot, Nx))
+            d.x  = RollingArray((K_plot, len(p.dims)))
             h.x  = []
-        if True:
-            d.y  = RollingArray((K_plot, Ny))
+        if p.obs_inds:
+            d.y  = RollingArray((K_plot, len(p.dims)))
             h.y  = []
         if E is not None:
-            d.E  = RollingArray((K_plot, len(E), Nx))
+            d.E  = RollingArray((K_plot, len(E), len(p.dims)))
             h.E  = []
         if P is not None:
-            d.mu = RollingArray((K_plot, Nx))
+            d.mu = RollingArray((K_plot, len(p.dims)))
             h.mu = []
         if P is not None:
-            d.s  = RollingArray((K_plot, 2, Nx))
+            d.s  = RollingArray((K_plot, 2, len(p.dims)))
             h.s  = []
 
         # Plot (invisible coz everything here is nan, for the moment).
-        for ix, (_m, iy, ax) in enumerate(zip(DimsX, DimsY, axs)):
+        for ix, ax in zip(p.dims, axs):
             if True:
                 h.x  += ax.plot(d.t, d.x[:, ix], 'k')
[…]
