### 2023-05-18 8801573
Define sectors for "double" Lorenz63
