+++ examples/time-dep-obs-operator.py
@@ -17,19 +17,19 @@ from dapper.mods.Lorenz63 import LPs, Tplot, dxdt
 tseq = modelling.Chronology(0.01, dko=25, Ko=1000, Tplot=Tplot, BurnIn=4*Tplot)
 
 
-def dxdt2(x):
+# Define a "double" Lorenz-63 system where each "half" evolves independently.
+def dxdt_double(x):
     x1 = x[..., :3]
     x2 = x[..., 3:]
     f1 = dxdt(x1)
-    f2 = dxdt(x2)
+    f2 = dxdt(10 * x2) / 10
     return np.concatenate([f1, f2], -1)
 
 
 Nx = 6
-
 Dyn = {
     'M': Nx,
-    'model': modelling.with_rk4(dxdt2, autonom=True),
+    'model': modelling.with_rk4(dxdt_double, autonom=True),
     'noise': 0,
 }
 
@@ -49,7 +49,8 @@ jj2 = 3 + np.arange(3)
 obs1 = modelling.Operator(**modelling.partial_Id_Obs(Nx, jj1), noise=1)
 obs2 = modelling.Operator(**modelling.partial_Id_Obs(Nx, jj2), noise=1)
 
-HMM = modelling.HiddenMarkovModel(Dyn, dict(time_dependent=Obs), tseq, X0)
+HMM = modelling.HiddenMarkovModel(Dyn, dict(time_dependent=Obs), tseq, X0,
+                                  sectors={'slow': jj1, 'fast': jj2})
 
 HMM.liveplotters = LPs(jj=lambda ko: jj1 if ko % 2 else jj2, params=dict())
 
@@ -66,6 +67,5 @@ xp.assimilate(HMM, xx, yy, liveplots=True)
 
 # print(xp.stats)  # ⇒ long printout
 xp.stats.average_in_time()
-
 # print(xp.avrgs)  # ⇒ long printout
-print(xp.avrgs.tabulate(['rmse.a', 'rmv.a']))
+print(xp.avrgs.tabulate(["rmse.a", "rmse.slow.a", "rmse.fast.a"]))
