+++ setup.py
@@ -24,7 +24,7 @@ INSTALL_REQUIRES = [
     'ipdb',
     'ipython>=5.1',
     # 'jedi<0.18',  # ipython/issues/12740
-    'tornado==6.1',  # 6.2 breaks Jupyter plots (tested on local Mac, Linux)
+    'tornado==6.3.2',  # 6.2 breaks Jupyter plots (tested on local Mac, Linux)
     # (pip says google-colab requires tornado~=6.2 but it still works)
     'matplotlib~=3.7',
     'mpl-tools==0.2.50',
