### 2023-05-26 71832bd
Fix check for obs_inds being ndarray
