### 2023-05-26 45c624c
Minor fixes to ipynb's
