### 2023-06-21 6741dc3
README: update limitations
