### 2023-08-02 #5150 @Shreyas911
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?

I get the following error when I create the conda environment dapper-env on our Linux machine for Python 3.16 as suggested in your README.

`Traceback (most recent call last):
  File "/home/shreyas/DAPPER/examples/basic_1.py", line 10, in <module>
    from mpl_tools import is_notebook_or_qt as nb
ModuleNotFoundError: No module named 'mpl_tools'`

I tried my own conda environments with Python 3.10 or above and still had the same error. I guess mpl_tools needs to be specifically added to the environment

> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?

I will verify these soon if I can get this `mpl_tools` issue sorted.

> ### Documentation
> * [ ]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The target audience is not explicitly specified in the documentation, I think.

> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally, these should be handled with an automated package management solution.

A list of dependencies should be stated, including `conda/mamba` along with a minimum version and maybe `mpl_tools` if that is to be separately added.

> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This is not explicitly stated. I do see a dev section, however, it would be great to include in the README.md file explicitly how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

This section is missing from the paper. I do see it in the README, so maybe add that to the paper as well.

> * [ ]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

A few minor suggestions -
* In the Statement of Need section, it is confusing to read (a) (b) (a) (b) (c) (d) (c) (d). Maybe change one of them to numbers or some other formatting.
* In the Acknowledgements sections, end the paragraph with a full stop instead of a comma.
