### 2023-08-03 35a4376
Pin patch py version
Hopefully fixes #106
