### 2023-08-03 1dd9e02
Try pinning Ubuntu instead
Revert
