### 2023-08-03 64b6b49
Try debuggin CI
Run artefact upload also when test fail
