### 2023-08-03 5020503
Create rng for DAPPER instead global RandomState
This upgrade was always due,
but also prompted by #106
