+++ dapper/tools/matrices.py
@@ -7,6 +7,7 @@ import scipy.linalg as sla
 from numpy import ones, sqrt, zeros
 
 from dapper.tools.linalg import mrdiv, svd0, truncate_rank
+from dapper.tools.seeding import rng
 
 
 class lazy_property:
@@ -32,7 +33,7 @@ class lazy_property:
 def randcov(M):
     """(Makeshift) random cov mat."""
     N = int(np.ceil(2+M**1.2))
-    E = np.random.randn(N, M)
+    E = rng.standard_normal((N, M))
     return E.T @ E
 
 
@@ -47,7 +48,7 @@ def genOG(M):
     """Generate random orthonormal matrix."""
     # TODO 5: This (using Householder) is (slightly?) wrong,
     # as per section 4 of mezzadri2006generate.
-    Q, R = sla.qr(np.random.randn(M, M))
+    Q, R = sla.qr(rng.standard_normal((M, M)))
     for i in range(M):
         if R[i, i] < 0:
             Q[:, i] = -Q[:, i]
