+++ dapper/da_methods/ensemble.py
@@ -1,7 +1,6 @@
 """The EnKF and other ensemble-based methods."""
 
 import numpy as np
-import numpy.random as rnd
 import scipy.linalg as sla
 from numpy import diag, eye, sqrt, zeros
 
@@ -11,6 +10,7 @@ from dapper.tools.linalg import mldiv, mrdiv, pad0, svd0, svdi, tinv, tsvd
 from dapper.tools.matrices import funm_psd, genOG_1
 from dapper.tools.progressbar import progbar
 from dapper.tools.randvars import GaussRV
+from dapper.tools.seeding import rng
 
 from . import da_method
 
@@ -161,10 +161,10 @@ def EnKF_analysis(E, Eo, hnoise, y, upd_a, stats=None, ko=None):
                         v     = v - mult[0]*Yj # noqa
                         v    /= sqrt(v@v)
                     Zj  = v*sqrt(N1)  # Standardized perturbation along v
-                    Zj *= np.sign(rnd.rand()-0.5)  # Random sign
+                    Zj *= np.sign(rng.standard_normal() - 0.5)  # Random sign
                 else:
                     # The usual stochastic perturbations.
-                    Zj = mean0(rnd.randn(N))  # Un-coloured noise
+                    Zj = mean0(rng.standard_normal(N))  # Un-coloured noise
                     if 'Var1' in upd_a:
                         Zj *= sqrt(N/(Zj@Zj))
 
@@ -319,7 +319,7 @@ def add_noise(E, dt, noise, method):
         E, _, Qa12 = sqrt_core()
         if N <= Nx:
             Z  = Q12 - A.T@Qa12
-            E += sqrt(dt)*(Z@rnd.randn(Z.shape[1], N)).T
+            E += sqrt(dt)*(Z@rng.standard_normal((Z.shape[1], N))).T
 
     elif method == 'Sqrt-Dep':
         E, T, Qa12 = sqrt_core()
@@ -334,7 +334,7 @@ def add_noise(E, dt, noise, method):
             Z      = Q12 - Q_hat12
             D_hat  = A.T@(T-eye(N))
             Xi_hat = Q_hat12_inv @ D_hat
-            Xi_til = (eye(rQ) - Q_hat12_proj)@rnd.randn(rQ, N)
+            Xi_til = (eye(rQ) - Q_hat12_proj)@rng.standard_normal((rQ, N))
             D_til  = Z@(Xi_hat + sqrt(dt)*Xi_til)
             E     += D_til.T
 
@@ -468,7 +468,7 @@ def serial_inds(upd_a, y, cvR, A):
             dC = np.sum(A*A, 0)/(N-1)
         inds = np.argsort(dC)
     else:  # Default: random ordering
-        inds = rnd.permutation(len(y))
+        inds = rng.permutation(len(y))
     return inds
 
 
