+++ dapper/da_methods/particle.py
@@ -1,12 +1,12 @@
 """Weight- & resampling-based DA methods."""
 
 import numpy as np
-import numpy.random as rnd
 
 from dapper.stats import unbias_var, weight_degeneracy
 from dapper.tools.linalg import mldiv, mrdiv, pad0, svd0, tinv
 from dapper.tools.matrices import chol_reduce, funm_psd
 from dapper.tools.progressbar import progbar
+from dapper.tools.seeding import rng
 
 from . import da_method
 
@@ -61,7 +61,7 @@ class PartFilt:
         for k, ko, t, dt in progbar(HMM.tseq.ticker):
             E = HMM.Dyn(E, t-dt, dt)
             if HMM.Dyn.noise.C != 0:
-                D  = rnd.randn(N, Nx)
+                D  = rng.standard_normal((N, Nx))
                 E += np.sqrt(dt*self.qroot)*(D@HMM.Dyn.noise.C.Right)
 
                 if self.qroot != 1.0:
@@ -116,7 +116,7 @@ class OptPF:
         for k, ko, t, dt in progbar(HMM.tseq.ticker):
             E = HMM.Dyn(E, t-dt, dt)
             if HMM.Dyn.noise.C != 0:
-                E += np.sqrt(dt)*(rnd.randn(N, Nx)@HMM.Dyn.noise.C.Right)
+                E += np.sqrt(dt)*(rng.standard_normal((N, Nx))@HMM.Dyn.noise.C.Right)
 
             if ko is not None:
                 self.stats.assess(k, ko, 'f', E=E, w=w)
@@ -188,7 +188,7 @@ class PFa:
         for k, ko, t, dt in progbar(HMM.tseq.ticker):
             E = HMM.Dyn(E, t-dt, dt)
             if HMM.Dyn.noise.C != 0:
-                D  = rnd.randn(N, Nx)
+                D  = rng.standard_normal((N, Nx))
                 E += np.sqrt(dt*self.qroot)*(D@HMM.Dyn.noise.C.Right)
 
                 if self.qroot != 1.0:
@@ -262,7 +262,7 @@ class PFxN_EnKF:
         for k, ko, t, dt in progbar(HMM.tseq.ticker):
             E = HMM.Dyn(E, t-dt, dt)
             if HMM.Dyn.noise.C != 0:
-                E += np.sqrt(dt)*(rnd.randn(N, Nx)@HMM.Dyn.noise.C.Right)
+                E += np.sqrt(dt)*(rng.standard_normal((N, Nx))@HMM.Dyn.noise.C.Right)
 
             if ko is not None:
                 self.stats.assess(k, ko, 'f', E=E, w=w)
@@ -290,7 +290,7 @@ class PFxN_EnKF:
                         Pa      = Aw.T@Aw - KG@Yw.T@Aw
                         P_cholU = funm_psd(Pa, np.sqrt)
                         if DD is None or not self.re_use:
-                            DD    = rnd.randn(N*xN, Nx)
+                            DD    = rng.standard_normal((N*xN, Nx))
                             chi2  = np.sum(DD**2, axis=1) * Nx/N
                             log_q = -0.5 * chi2
                     else:
@@ -303,7 +303,7 @@ class PFxN_EnKF:
                         # and compute log(q(x))
                         if DD is None or not self.re_use:
                             rnk   = min(Nx, N-1)
-                            DD    = rnd.randn(N*xN, N)
+                            DD    = rng.standard_normal((N*xN, N))
                             chi2  = np.sum(DD**2, axis=1) * rnk/N
                             log_q = -0.5 * chi2
                         # NB: the DoF_linalg/DoF_stoch correction
@@ -378,7 +378,7 @@ class PFxN:
[…]
