### 2023-08-03 deb07f7
Fix svd orientation issue (scipy 1.10 --> 1.11)
