+++ tests/test_example_2.py
@@ -77,11 +77,11 @@ L63_old = """
  8  EnKF          0.95  PertObs   500  False                   |     0.5719 ±0.0854     1.5111 ±0.3688     0.8907 ±0.1971
  9  EnKF_N        1                10  True      1             |     0.6230 ±0.0676     1.6784 ±0.3686     0.9912 ±0.1970
 10  iEnKS         1.02  Sqrt       10  True                    |     0.3581 ±0.1094     0.9190 ±0.2488     0.2589 ±0.0792
-11  PartFilt                      100               2.4  0.3   |     0.4814 ±0.0758     1.0920 ±0.2710     0.6794 ±0.1755
-12  PartFilt                      800               0.9  0.2   |     0.3778 ±0.1020     0.8797 ±0.2251     0.5502 ±0.1760
-13  PartFilt                     4000               0.7  0.05  |     0.3452 ±0.1099     0.8395 ±0.2293     0.5220 ±0.1811
-14  PFxN                           30         1000       0.2   |     0.4278 ±0.0896     1.4632 ±0.4063     0.8819 ±0.2508
-15  OptPF                         100               0.7  0.3   |     0.6125 ±0.1085     1.5794 ±0.4536     0.9871 ±0.3130
+11  PartFilt                      100               2.4  0.3   |     0.4665 ±0.0932     1.1912 ±0.3142     0.7288 ±0.1990
+12  PartFilt                      800               0.9  0.2   |     0.3630 ±0.0979     0.8823 ±0.2273     0.5529 ±0.1766
+13  PartFilt                     4000               0.7  0.05  |     0.3468 ±0.1096     0.8412 ±0.2295     0.5231 ±0.1808
+14  PFxN                           30         1000       0.2   |     0.5391 ±0.1022     1.4199 ±0.3777     0.8734 ±0.2618
+15  OptPF                         100               0.7  0.3   |     0.6512 ±0.1232     1.6482 ±0.4774     1.0154 ±0.3177
 16  EnKS          1     Serial     30  False                   |     0.5997 ±0.0829     1.5564 ±0.3523     0.4004 ±0.0640
 17  EnRTS         1     Serial     30  False                   |     0.5997 ±0.0829     1.5564 ±0.3523     0.2485 ±0.0637
 """[1:-1].splitlines(True)
