### 2024-01-17 #5150 @sara-02
Hey @patnr I was able to test out random examples from the examples directory but I faced two issues.1
1. In the `pytest tests` the HMM test is failing.
2. The last example `https://github.com/nansencenter/DAPPER/blob/master/examples/time-dep-obs-operator.py` got stuck. Could be an issue from running it from Python terminal or a RAM issue? I am on Ubuntu 22 with 8gb of RAM.
Should I be reporting any of these issues on your repo?
Rest all looks good.
