### 2024-01-18 #5150 @patnr
Hi @sara-02 and thank you for the feedback.

> 1. In the pytest tests the HMM test is failing.

I'm guessing it's running some tests that it should not (should only run if the Fortran code has been compiled). It could be that this is already fixed, so if you're on an old commit then could you please try pulling the master branch ? If it's still failing then could you please copy-paste the full error message/output/traceback from the failing test?

> 2. The last example https://github.com/nansencenter/DAPPER/blob/master/examples/time-dep-obs-operator.py got stuck.

This could again be because of an old commit. Otherwise I suspect that it's a matplotlib issue. Do you know where it hangs? Or do you get an error?

> Should I be reporting any of these issues on your repo?

If you think that this will become a lengthy thread then please do. Otherwise I guess it's fine to continue the discussion here.
