### 2024-02-26 #5150 @sara-02
> @sara-02 Ah no, that is definitely recent enough!

I reinstalled the library and this time I did not do the install with `.dev` and the example issue is fixed, but test cases are still failing.
