+++ dapper/mods/QG/f90/README.md
@@ -26,52 +26,56 @@ Licence reproduced here.
 Compile as follows...
 
 ### This `f90` directory
-contains Fortran90 code for the QG model for building
+
+contains Fortran-90 code for the QG model for building
 
 1. a python extension module `py_mod`
 2. a standalone program
 
 DAPPER only requires `py_mod` (also to generate initial sample).
-Both require an f90 compiler (tested with g95 and gfortran).  
+Both require an `f90` compiler (tested with `g95` and `gfortran`).  
 In addition  `py_mod` requires `f2py`, while `qg` requires `netcdf` libraries.
 
 
-### For DAPPER,
-To build `py_mod`, ensure that both `gcc` and `gfortran` are installed. For example, with `conda`:
-    
-    $ conda install -c conda-forge gcc
-    $ conda install -c conda-forge gfortran
-    
+### For DAPPER
+
+To build `py_mod`, ensure that both `gcc` and `gfortran` are installed.
+For example, with `conda`:
+
+    conda install -c conda-forge gcc
+    conda install -c conda-forge gfortran
+
 Then run:
 
-    $ cd dapper/mods/QG/f90
-    $ rm -rf py_mod.cpython-* __pycache__
-    $ f2py -c utils.f90 parameters.f90 helmholtz.f90 calc.f90 qgflux.f90 qgstep.f90 interface.f90 -m py_mod
+    cd dapper/mods/QG/f90
+    rm -rf py_mod.cpython-* __pycache__
+    f2py -c utils.f90 parameters.f90 helmholtz.f90 calc.f90 qgflux.f90 qgstep.f90 interface.f90 -m py_mod
 
 ### For the standalone executable `qg`
+
 (not required for DAPPER), adapt the `Makefile` to your system, and run
 
-    $ make qg
+    make qg
 
 Example: here's how I compiled the standalone on my Mac:
 
-- get netcdf: `brew install netcdf --with-fortran`
-- In makefile, changed to:
-      FC = gfortran-5
-      NCLIB = /usr/local/lib/libnetcdff.dylib /usr/local/lib/libnetcdf.dylib
-- Matlab has a new netcdf interface. Therefore, qgread.m should use
-      % ncdisp(fname)
-      ncid = netcdf.open(fname);
-      psi  = permute( netcdf.getVar(ncid, netcdf.inqVarID(ncid,'psi')), [3 2 1]);
-      q    = permute( netcdf.getVar(ncid, netcdf.inqVarID(ncid,'q'))  , [3 2 1]);
-      t    = netcdf.getVar(ncid, netcdf.inqVarID(ncid,'t'));
+    - get netcdf: `brew install netcdf --with-fortran`
+    - In makefile, changed to:
+          FC = gfortran-5
+          NCLIB = /usr/local/lib/libnetcdff.dylib /usr/local/lib/libnetcdf.dylib
+    - Matlab has a new netcdf interface. Therefore, qgread.m should use
+          % ncdisp(fname)
+          ncid = netcdf.open(fname);
+          psi  = permute( netcdf.getVar(ncid, netcdf.inqVarID(ncid,'psi')), [3 2 1]);
+          q    = permute( netcdf.getVar(ncid, netcdf.inqVarID(ncid,'q'))  , [3 2 1]);
+          t    = netcdf.getVar(ncid, netcdf.inqVarID(ncid,'t'));
 
-### Changelog since Sakov's enkf-matlab:
+### Changelog since Sakov's enkf-matlab
 
-- Rm Matlab interface funcs: qgplay.m qgplot.m qgread.m mexcmd.m qgstep_mex.f90 mexf90.f90
-- Modified makefile, as described above.
-- In parameters.f90:
-    - Capitalized m,n
[…]
