### 2024-02-26 408a9ff
Imp. error can arise both at initial import and at run
Therefore, wrap entire test in try/except ModuleNotFoundError

Fixes #109
