### 2024-02-26 a22e36d
Set Zenodo DOI for v1.6.0 release
