### 2024-02-26 #5150 @patnr
Thank you for the guidelines.

**[v1.6.0](https://github.com/nansencenter/DAPPER/releases/tag/v1.6.0)**
**[10.5281/zenodo.10710355](https://zenodo.org/records/10710355)**
