### 2024-02-28 2a99a4c
Dont suggest [dev] install by default
