### 2024-02-28 17dc643
Improve docs for defining new model
