+++ dapper/mods/__init__.py
@@ -1,4 +1,4 @@
-"""Models included with DAPPER.
+"""Contains models included with DAPPER.
 
 .. include:: ./README.md
 """
@@ -25,14 +25,30 @@ from .utils import Id_Obs, ens_compatible, linspace_int, partial_Id_Obs
 
 
 class HiddenMarkovModel(struct_tools.NicePrint):
-    """Container for a Hidden Markov Model (HMM).
+    """Container class (with some embellishments) for a Hidden Markov Model (HMM).
 
-    This should contain the details necessary to run synthetic DA experiments,
+    Should contain the details necessary to run synthetic DA experiments,
     also known as "twin experiment", or OSSE (observing system simulation experiment).
     The synthetic truth and observations may then be obtained by running
     `HiddenMarkovModel.simulate`.
 
-    See scripts in examples for more details.
+    .. warning::
+      Each model included with DAPPER comes with several examples
+      of model settings from the literature.
+      See, for example, `dapper.mods.Lorenz63.sakov2012`.
+      These configurations do not necessarily hold a very high programming standard,
+      as they may have been whipped up at short notice to replicate some experiments,
+      and are not intended for re-use.
+
+      Nevertheless, sometimes they are re-used by another configuration script,
+      leading to a major gotcha/pitfall: changes made to the imported `HMM` (or
+      the model's module itself) also impact the original object (since they
+      are mutable and thereby referenced).  This *usually* isn't an issue, since
+      one rarely imports two/more separate configurations. However, the test suite
+      imports all configurations, which might then unintentionally interact.
+      To avoid this, you should use the `copy` method of the `HMM`
+      before making any changes to it.
+
 
     Parameters
     ----------
